requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
decorator==4.4.2
distlib==0.3.1
Django==3.1.4
django-crispy-forms==1.10.0
djangorestframework==3.12.2
filelock==3.0.12
gitdb==4.0.5
GitPython==3.1.12
idna==2.10
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
jupyter-client==6.1.7
jupyter-core==4.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.8
Pygments==2.7.3
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2020.5
pywin32==300
pyzmq==20.0.0
requests==2.25.1
six==1.15.0
smmap==3.0.5
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.2
virtualenv==20.2.2
wcwidth==0.2.5
wrapt==1.12.1
